requirements.txt
appdirs==1.4.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
distlib==0.3.1
et-xmlfile==1.1.0
filelock==3.0.12
future==0.18.2
idna==2.10
kiwisolver==1.3.1
matplotlib==3.4.2
numexpr==2.7.3
numpy==1.20.2
openpyxl==3.0.7
pandas==1.2.4
pandastable==0.12.2.post1
pathvalidate==2.4.1
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytest==6.2.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
tk==0.1.0
urllib3==1.26.4
virtualenv==20.4.6
xlrd==2.0.1